move useComponentCreate hook into its own component
