+++ designer/client/components/ComponentCreate/useComponentCreate.tsx
@@ -1,12 +1,12 @@
-import React, { useEffect, useContext, useState, useLayoutEffect } from "react";
-import { ComponentDef } from "@xgovformbuilder/model";
-
-import { i18n } from "../../i18n";
-import { ErrorSummary } from "../../error-summary";
+import {
+  useEffect,
+  useContext,
+  useState,
+  useLayoutEffect,
+  FormEvent,
+} from "react";
+import { ComponentDef, Page } from "@xgovformbuilder/model";
 import { hasValidationErrors } from "../../validations";
-import ComponentTypeEdit from "../../ComponentTypeEdit";
-import { ComponentCreateList } from "./ComponentCreateList";
-import { BackLink } from "../BackLink";
 import "./ComponentCreate.scss";
 import { Actions } from "../../reducers/component/types";
 import { DataContext } from "../../context";
@@ -14,30 +14,52 @@ import { ComponentContext } from "../../reducers/component/componentReducer";
 import { addComponent } from "../../data";
 import logger from "../../plugins/logger";
 
-function useComponentCreate(props) {
+export const useComponentCreate = (props) => {
+  const [renderTypeEdit, setRenderTypeEdit] = useState<boolean>(false);
   const { data, save } = useContext(DataContext);
   const { state, dispatch } = useContext(ComponentContext);
-  const { selectedComponent, errors = {}, hasValidated } = state;
+  const { selectedComponent, errors = [], hasValidated } = state;
   const { page, toggleAddComponent = () => {} } = props;
 
   const [isSaving, setIsSaving] = useState(false);
   const hasErrors = hasValidationErrors(errors);
 
   useEffect(() => {
+    // render in the next re-paint to allow the DOM to reflow without the list
+    // thus resetting the Flyout wrapper scrolling position
+    // This is a quick work around the bug in small screens
+    // where once user scrolls down the components list and selects one of the bottom components
+    // then the component edit screen renders already scrolled to the bottom
+    let isMounted = true;
+
+    if (selectedComponent?.type) {
+      window.requestAnimationFrame(() => {
+        if (isMounted) setRenderTypeEdit(true);
+      });
+    } else {
+      setRenderTypeEdit(false);
+    }
+
+    return () => {
+      isMounted = false;
+    };
+  }, [selectedComponent?.type]);
+
+  useEffect(() => {
     dispatch({ type: Actions.SET_PAGE, payload: page.path });
-  }, [dispatch, page.path]);
+  }, [page.path]);
 
   useLayoutEffect(() => {
     if (hasValidated && !hasErrors) {
       handleSubmit()
         .then()
         .catch((err) => {
-          logger.error("useComponentCreate", err);
+          logger.error("ComponentCreate", err);
         });
     }
   }, [hasValidated, hasErrors]);
 
-  const handleSubmit = async (e?: Event) => {
+  const handleSubmit = async (e?: FormEvent<HTMLFormElement>) => {
     e?.preventDefault();
 
     if (!hasValidated) {
@@ -50,14 +72,24 @@ function useComponentCreate(props) {
     }
 
     setIsSaving(true);
-    const { isNew, ...selectedComponent } = state.selectedComponent;
-    const updatedData = addComponent(data, page.path, selectedComponent);
+    const { selectedComponent } = state;
+    const updatedData = addComponent(
+      data,
+      (page as Page).path,
+      selectedComponent
+    );
+
     await save(updatedData);
     toggleAddComponent();
   };
 
   const handleTypeChange = (component: ComponentDef) => {
-    dispatch({ type: Actions.EDIT_TYPE, payload: component.type });
+    dispatch({
+      type: Actions.EDIT_TYPE,
+      payload: {
+        type: component.type,
+      },
+    });
   };
 
   const reset = (e) => {
@@ -73,51 +105,6 @@ function useComponentCreate(props) {
     component: selectedComponent,
     isSaving,
     reset,
+    renderTypeEdit,
   };
-}
-
-export function ComponentCreate(props) {
-  const {
-    handleSubmit,
-    handleTypeChange,
-    reset,
-    hasErrors,
-    errors,
-    component,
-    isSaving,
-  } = useComponentCreate(props);
-  const type = component?.type;
-
-  return (
-    <div className="component-create">
-      {hasErrors && <ErrorSummary errorList={errors} />}
-      {!type && <h4 className="govuk-heading-m">{i18n("Create component")}</h4>}
-      {type && (
-        <>
-          <BackLink onClick={reset}>
-            {i18n("Back to create component list")}
-          </BackLink>
-          <h4 className="govuk-heading-m">
-            {component?.["title"]} {i18n("component")}
-          </h4>
-        </>
-      )}
-      <form onSubmit={handleSubmit}>
-        <div className="govuk-form-group">
-          <label className="govuk-label govuk-label--s" htmlFor="type">
-            Type
-          </label>
-          {!type && (
-            <ComponentCreateList onSelectComponent={handleTypeChange} />
-          )}
-        </div>
-        {type && <ComponentTypeEdit />}
-        <button type="submit" className="govuk-button" disabled={isSaving}>
-          Save
-        </button>
-      </form>
-    </div>
-  );
-}
-
-export default ComponentCreate;
+};
